refactor(ng-add): Update schematics #3528
- Separate utility functions
- Update tsconfig.json:outDir
- Separate Options iterface
