+++ projects/igniteui-angular/schematics/interfaces/options.ts
@@ -1,0 +1,4 @@
+export interface Options {
+    [key: string]: any;
+    polyfills: boolean;
+}
